Add StreamingInstructionParser
To be able to run larger experiments we need to stream in our traces as
they become the bottleneck on the amount of runs we can run in
parallel.  The StreamingInstructionParser does not work for Belady's MIN
as due to performance reasons it needs to pre-processes the entire
trace.

By using the StreamingInstructionParser the amount of memory used for
each experiment seems to be <700MB.